backend/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
deprecation==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
modzy-sdk==0.6.0
packaging==21.3
pyparsing==3.0.6
python-dotenv==0.19.2
pytube==11.0.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
youtube-transcript-api==0.4.2
